Rename useGraphNodeSubscribe to useGraphNodeSnapshot
