+++ packages/preact-graph-state/src/hooks/useGraphNodeSnapshotBase.tsx
