+++ packages/preact-graph-state/src/index.ts
@@ -29,7 +29,7 @@ export { default as useGraphNodeValue } from './hooks/useGraphNodeValue';
 export { default as useGraphNodeState } from './hooks/useGraphNodeState';
 export { default as useGraphNodeDispatch } from './hooks/useGraphNodeDispatch';
 export { default as useGraphNodeResource } from './hooks/useGraphNodeResource';
-export { default as useGraphNodeSubscribe } from './hooks/useGraphNodeSubscribe';
+export { default as useGraphNodeSnapshot } from './hooks/useGraphNodeSnapshot';
 export { default as GraphDomain } from './GraphDomain';
 
 export { GraphNodeDispatch } from './hooks/useGraphNodeDispatchBase';
